+++ frontend/pages/bewerten.tsx
@@ -1,11 +1,12 @@
-// Diese Datei repräsentiert die Seite /bewerten.tsx mit erweitertem UX-Hinweis bei Pflichtfeldfehlern
+// frontend/pages/bewerten.tsx - Modernes Wizard Design
 
 import Head from "next/head";
 import Link from "next/link";
-import React, { useState } from "react";
+import Image from "next/image";
+import React, { useState, useEffect } from "react";
 import { error } from "@/lib/log";
-import Layout from "@/components/Layout"; // Footer via Layout integriert
-
+import Layout from "@/components/Layout";
+import { Star, CheckCircle, Lock, Zap, ArrowRight, ArrowLeft } from "lucide-react";
 
 interface FormState {
   rasse: string;
@@ -37,86 +38,220 @@ const initialForm: FormState = {
   verwendungszweck: "",
 };
 
-const fields = [
-  { name: "rasse", label: "Rasse", required: true },
-  { name: "alter", label: "Alter (Jahre)", type: "number", required: true },
+// Field Interface
+interface FormField {
+  name: string;
+  label: string;
+  required?: boolean;
+  type?: string;
+  placeholder?: string;
+  fullWidth?: boolean;
+  halfWidth?: boolean;
+  options?: string[];
+}
+
+// Step Interface
+interface StepData {
+  id: number;
+  title: string;
+  subtitle: string;
+  description: string;
+  icon: string;
+  iconBg: string;
+  fields: FormField[];
+}
+
+// Step-Konfiguration (Wizard-Fortschritt)
+const stepData: StepData[] = [
   {
-    name: "geschlecht",
-    label: "Geschlecht",
-    type: "select",
-    required: true,
-    options: ["Stute", "Wallach", "Hengst"],
+    id: 1,
+    title: "Grunddaten",
+    subtitle: "Grunddaten deines Pferdes",
+    description: "Erzähle uns die wichtigsten Daten zu deinem Pferd",
+    icon: "🐎",
+    iconBg: "bg-amber-100",
+    fields: [
+      { 
+        name: "rasse", 
+        label: "Rasse", 
+        required: true, 
+        placeholder: "z.B. Deutsches Sportpferd, Hannoveraner, Oldenburger",
+        fullWidth: true
+      },
+      { 
+        name: "alter", 
+        label: "Alter (Jahre)", 
+        type: "number", 
+        required: true, 
+        placeholder: "z.B. 8",
+        halfWidth: true
+      },
+      { 
+        name: "geschlecht", 
+        label: "Geschlecht", 
+        type: "select", 
+        required: true, 
+        options: ["Stute", "Wallach", "Hengst"],
+        halfWidth: true
+      },
+      { 
+        name: "stockmass", 
+        label: "Stockmaß (cm)", 
+        type: "number", 
+        required: true, 
+        placeholder: "z.B. 168",
+        fullWidth: true
+      },
+    ]
   },
   {
-    name: "abstammung",
-    label: "Abstammung (Vater x Muttervater)",
-    required: true,
-    placeholder: "z. B. Cornet Obolensky x Contender",
+    id: 2,
+    title: "Details",
+    subtitle: "Abstammung & Ausbildung",
+    description: "Details zur Leistung und Ausbildung deines Pferdes",
+    icon: "🏆",
+    iconBg: "bg-blue-100",
+    fields: [
+      { 
+        name: "abstammung", 
+        label: "Abstammung (Vater x Muttervater)", 
+        required: true, 
+        placeholder: "z.B. Cornet Obolensky x Contender",
+        fullWidth: true
+      },
+      { 
+        name: "ausbildung", 
+        label: "Ausbildungsstand", 
+        required: true, 
+        placeholder: "z.B. A-Dressur, Springpferde L, angeritten",
+        fullWidth: true
+      },
+      { 
+        name: "verwendungszweck", 
+        label: "Verwendungszweck", 
+        required: false,
+        placeholder: "z.B. Freizeit, Turnier, Zucht",
+        fullWidth: true
+      },
+    ]
   },
   {
-    name: "stockmass",
-    label: "Stockmaß (cm)",
-    type: "number",
-    required: true,
+    id: 3,
+    title: "Zusatzinfos",
+    subtitle: "Zusätzliche Informationen",
+    description: "Optionale Details für eine noch genauere Bewertung",
+    icon: "📋",
+    iconBg: "bg-green-100",
+    fields: [
+      { 
+        name: "aku", 
+        label: "Gesundheitsstatus / AKU-Bericht", 
+        required: false,
+        placeholder: "z.B. AKU ohne Befund, kleine Befunde",
+        fullWidth: true
+      },
+      { 
+        name: "erfolge", 
+        label: "Erfolge", 
+        required: false,
+        placeholder: "z.B. L-Platzierungen Springen, A-Dressur gewonnen",
+        fullWidth: true
+      },
+      { 
+        name: "farbe", 
+        label: "Farbe", 
+        required: false,
+        placeholder: "z.B. Fuchs, Rappe, Brauner",
+        halfWidth: true
+      },
+      { 
+        name: "standort", 
+        label: "Standort", 
+        required: false,
+        placeholder: "z.B. München oder PLZ",
+        halfWidth: true
+      },
+      { 
+        name: "zuechter", 
+        label: "Züchter", 
+        required: false,
+        placeholder: "z.B. Privat oder Zuchtbetrieb XYZ",
+        fullWidth: true
+      },
+    ]
   },
   {
-    name: "ausbildung",
-    label: "Ausbildungsstand",
-    required: true,
-    placeholder: "z. B. A-Dressur, Springpferde L, angeritten",
-  },
-  {
-    name: "aku",
-    label: "Gesundheitsstatus / AKU-Bericht",
-    placeholder: "z. B. AKU ohne Befund",
-  },
-  {
-    name: "erfolge",
-    label: "Erfolge",
-    placeholder: "z. B. L-Platzierungen Springen, A-Dressur gewonnen",
-  },
-  {
-    name: "farbe",
-    label: "Farbe",
-    placeholder: "z. B. Fuchs, Rappe, Brauner",
-  },
-  {
-    name: "zuechter",
-    label: "Züchter",
-    placeholder: "z. B. Privat oder Zuchtbetrieb XYZ",
-  },
-  {
-    name: "standort",
-    label: "Standort",
-    placeholder: "z. B. München oder PLZ",
-  },
-  {
-    name: "verwendungszweck",
-    label: "Verwendungszweck",
-    placeholder: "z. B. Freizeit, Turnier, Zucht",
-  },
+    id: 4,
+    title: "Bezahlung",
+    subtitle: "Analyse starten",
+    description: "Nur noch ein Klick zur professionellen Pferdebewertung",
+    icon: "💳",
+    iconBg: "bg-purple-100",
+    fields: []
+  }
 ];
 
-
-export default function Bewerten() {
+export default function TestBewPage() {
+  const [currentStep, setCurrentStep] = useState(1);
   const [form, setForm] = useState(initialForm);
   const [errors, setErrors] = useState<{ [key: string]: string }>({});
   const [loading, setLoading] = useState(false);
   const [consent, setConsent] = useState(false);
   const [submitted, setSubmitted] = useState(false);
 
+  // Formular bei Start wiederherstellen
+  useEffect(() => {
+    const savedForm = localStorage.getItem("bewertungForm");
+    if (savedForm) {
+      try {
+        const parsedForm = JSON.parse(savedForm);
+        setForm(parsedForm);
+        console.log("Formular aus localStorage wiederhergestellt");
+      } catch (e) {
+        console.warn("Fehler beim Wiederherstellen des Formulars:", e);
+      }
+    }
+  }, []);
+
+  // Formularwerte speichern bei jedem Change
+  useEffect(() => {
+    localStorage.setItem("bewertungForm", JSON.stringify(form));
+  }, [form]);
+
   const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
     setForm({ ...form, [e.target.name]: e.target.value });
+    // Clear error when user starts typing
+    if (errors[e.target.name]) {
+      setErrors({ ...errors, [e.target.name]: "" });
+    }
+  };
+
+  const validateStep = (step: number) => {
+    const currentStepData = stepData.find(s => s.id === step);
+    if (!currentStepData) return true;
+
+    const newErrors: { [key: string]: string } = {};
+    currentStepData.fields.forEach((field) => {
+      if (field.required && !form[field.name as keyof FormState]) {
+        newErrors[field.name] = "Pflichtfeld";
+      }
+    });
+
+    setErrors(newErrors);
+    return Object.keys(newErrors).length === 0;
+  };
+
+  const nextStep = () => {
+    if (validateStep(currentStep)) {
+      setCurrentStep(prev => Math.min(prev + 1, stepData.length));
+    }
+  };
+
+  const prevStep = () => {
+    setCurrentStep(prev => Math.max(prev - 1, 1));
   };
 
   const handleSubmit = async (e: React.FormEvent) => {
-    if (window.gtag) {
-      window.gtag("event", "start_bewertung", {
-        event_category: "Funnel",
-        event_label: "Formular abgeschickt",
-        value: 1,
-      });
-    }
     e.preventDefault();
     setSubmitted(true);
     setErrors({});
@@ -126,15 +261,26 @@ export default function Bewerten() {
       return;
     }
 
+    // Validate all required fields from all steps
     const newErrors: { [key: string]: string } = {};
-    fields.forEach((field) => {
-      if (field.required && !form[field.name as keyof FormState]) {
-        newErrors[field.name] = "Pflichtfeld";
-      }
+    stepData.slice(0, 3).forEach(step => {
+      step.fields.forEach((field) => {
+        if (field.required && !form[field.name as keyof FormState]) {
+          newErrors[field.name] = "Pflichtfeld";
+        }
+      });
     });
 
     if (Object.keys(newErrors).length > 0) {
       setErrors(newErrors);
+      // Go back to first step with errors
+      for (let i = 0; i < stepData.length - 1; i++) {
+        const stepFields = stepData[i].fields.map(f => f.name);
+        if (stepFields.some(field => newErrors[field])) {
+          setCurrentStep(i + 1);
+          break;
+        }
+      }
       return;
     }
 
@@ -148,6 +294,9 @@ export default function Bewerten() {
 
       if (res.ok) {
         const { url } = await res.json();
+        // Formular aus localStorage löschen bei erfolgreicher Einreichung
+        localStorage.removeItem("bewertungForm");
+        console.log("Formular erfolgreich eingereicht - localStorage geleert");
         window.location.href = url;
       } else {
         const { error } = await res.json();
@@ -165,152 +314,346 @@ export default function Bewerten() {
     }
   };
 
+  const currentStepData = stepData.find(s => s.id === currentStep);
+
   return (
     <Layout>
       <Head>
-  <title>Pferd bewerten & Pferdepreis ermitteln – in 2 Minuten zur fundierten Einschätzung | PferdeWert</title>
-  <meta name="description" content="Jetzt dein Pferd bewerten & den realistischen Marktwert online ermitteln – anonym, sicher & direkt als PDF. Ideal bei Pferdekauf & Pferdeverkauf." />
+        <title>Pferd bewerten & Pferdepreis ermitteln – in 2 Minuten zur fundierten Einschätzung | PferdeWert</title>
+        <meta name="description" content="Jetzt dein Pferd bewerten & den realistischen Marktwert online ermitteln – anonym, sicher & direkt als PDF. Ideal bei Pferdekauf & Pferdeverkauf." />
+        <link rel="canonical" href="https://pferdewert.de/bewerten" />
+      </Head>
 
-  {/* Open Graph Tags */}
-  <meta property="og:title" content="Pferd bewerten & Marktwert ermitteln – fundierte Einschätzung | PferdeWert" />
-  <meta property="og:description" content="Jetzt Pferd bewerten & Marktwert ermitteln – schnell, anonym & sicher. Ideal bei Pferdekauf und Pferdeverkauf." />
-  <meta property="og:type" content="website" />
-  <meta property="og:url" content="https://pferdewert.de/bewerten" />
-  <meta property="og:image" content="https://pferdewert.de/images/hero.webp" />
-
-  {/* Twitter Card */}
-  <meta name="twitter:card" content="summary_large_image" />
-  <meta name="twitter:title" content="Pferd bewerten & Marktwert ermitteln – fundierte Einschätzung | PferdeWert" />
-  <meta name="twitter:description" content="Jetzt Pferd bewerten & Marktwert ermitteln – schnell, anonym & sicher. Ideal bei Pferdekauf und Pferdeverkauf." />
-  <meta name="twitter:image" content="https://pferdewert.de/images/hero.webp" />
-
-  {/* Strukturierte Daten */}
-  <script type="application/ld+json">
-    {`
-    {
-      "@context": "https://schema.org",
-      "@type": "WebPage",
-      "name": "Pferd bewerten & Marktwert ermitteln – fundierte Einschätzung",
-      "description": "Jetzt Pferd bewerten & Marktwert ermitteln – schnell, anonym & sicher. Ideal bei Pferdekauf und Pferdeverkauf.",
-      "url": "https://pferdewert.de/bewerten"
-    }
-    `}
-  </script>
-  <link rel="canonical" href="https://pferdewert.de/bewerten" />
-</Head>
-
-
-      <main className="max-w-2xl mx-auto px-4 py-8">
-        <h1 className="text-3xl font-bold mb-2 text-center">Pferd analysieren lassen</h1>
-        <p className="text-sm text-gray-600 text-center mb-6">
-          🔒 Keine Anmeldung nötig – deine Eingaben bleiben anonym und werden nicht gespeichert.
-        </p>
-
-        <form onSubmit={handleSubmit} className="space-y-6">
-          {fields.map((field) => (
-            <div key={field.name}>
-              <label htmlFor={field.name} className="block font-medium mb-1">
-                {field.label}
-                {field.required && <span className="text-red-600 ml-1">*</span>}
-              </label>
-              {field.type === "select" ? (
-                <select
-                  id={field.name}
-                  name={field.name}
-                  value={form[field.name as keyof FormState]}
-                  onChange={handleChange}
-                  className={`w-full p-3 border rounded-xl transition ${
-                    errors[field.name] ? "border-red-500" : "border-gray-300"
-                  } focus:border-brand-accent focus:outline-none`}
-                >
-                  <option value="">Bitte wählen</option>
-                  {field.options?.map((option) => (
-                    <option key={option} value={option}>
-                      {option}
-                    </option>
-                  ))}
-                </select>
-              ) : (
-                <input
-                  id={field.name}
-                  name={field.name}
-                  type={field.type || "text"}
-                  value={form[field.name as keyof FormState]}
-                  onChange={handleChange}
-                  autoComplete="off"
-                  inputMode={field.type === "number" ? "numeric" : undefined}
-                  placeholder={field.placeholder}
-                  className={`w-full p-3 border rounded-xl transition ${
-                    errors[field.name] ? "border-red-500" : "border-gray-300"
-                  } focus:border-brand-accent focus:outline-none`}
-                />
-              )}
-              {errors[field.name] && (
-                <p className="mt-1 text-red-600 text-sm">{errors[field.name]}</p>
-              )}
+      {/* Hero-Bereich */}
+      <section className="bg-gradient-to-br from-[#fdf9f4] to-[#fef3c7] py-16 px-6 -mx-4 mb-8">
+        <div className="max-w-6xl mx-auto grid md:grid-cols-2 gap-10 items-center">
+          {/* Text-Bereich */}
+          <div>
+            <h1 className="text-4xl sm:text-5xl font-bold text-gray-900 mb-6">
+              Pferd analysieren
+            </h1>
+            
+            {/* Preisbanner */}
+            <div className="bg-gradient-to-r from-orange-100 to-yellow-100 border border-orange-200 p-4 rounded-xl shadow-sm mb-6">
+              <p className="text-lg font-semibold text-gray-800">
+                💥 Nur <span className="text-red-600 font-bold text-xl">4,90 €</span>
+                <span className="line-through text-gray-500 text-base ml-2">statt 39 €</span>
+              </p>
+              <p className="text-sm text-gray-600 mt-1">Für die ersten 100 Bewertungen!</p>
             </div>
-          ))}
 
-          <div className="text-sm text-gray-700">
-            <label className="flex items-start gap-2">
-              <input
-                type="checkbox"
-                checked={consent}
-                onChange={(e) => setConsent(e.target.checked)}
-                className="mt-1"
-                required
-              />
-              <span>
-                Ich stimme ausdrücklich zu, dass die Analyse sofort beginnt, und bestätige mein Erlöschen des Widerrufsrechts gemäß § 356 Abs. 5 BGB.
-              </span>
-            </label>
+            {/* Features */}
+            <div className="space-y-3 mb-6">
+              <div className="flex items-center gap-2 text-gray-700">
+                <Lock className="w-5 h-5 text-green-600" />
+                <span>🔒 Komplett anonym – keine Anmeldung nötig</span>
+              </div>
+              <div className="flex items-center gap-2 text-gray-700">
+                <Zap className="w-5 h-5 text-blue-600" />
+                <span>⚡ Ergebnis in unter 2 Minuten</span>
+              </div>
+              <div className="flex items-center gap-2 text-gray-700">
+                <CheckCircle className="w-5 h-5 text-amber-600" />
+                <span>📊 Detaillierte PDF-Analyse</span>
+              </div>
+            </div>
           </div>
 
-          {Object.keys(errors).some((key) => key !== "form") && submitted && (
-            <p className="text-red-600 text-center font-medium mt-2">
-              Bitte fülle alle markierten Pflichtfelder aus.
-            </p>
+          {/* Bild-Bereich */}
+          <div className="order-1 md:order-none">
+            <Image
+              src="/images/result.webp"
+              width={600}
+              height={400}
+              alt="Deutsches Sportpferd für KI-Pferdebewertung"
+              className="rounded-xl shadow-lg w-full h-auto"
+            />
+          </div>
+        </div>
+      </section>
+
+      {/* Wizard-Bereich */}
+      <section className="max-w-4xl mx-auto px-4 py-8">
+        {/* Step-Indikatoren */}
+        <div className="mb-8">
+          <div className="flex items-center justify-center space-x-8">
+            {stepData.map((step, index) => (
+              <div key={step.id} className="flex items-center">
+                <div className={`
+                  w-10 h-10 rounded-full flex items-center justify-center text-sm font-semibold transition-all
+                  ${currentStep >= step.id 
+                    ? 'bg-amber-600 text-white' 
+                    : 'bg-gray-200 text-gray-500'}
+                `}>
+                  {step.id}
+                </div>
+                <span className={`ml-2 text-sm font-medium hidden sm:block ${
+                  currentStep >= step.id ? 'text-amber-600' : 'text-gray-500'
+                }`}>
+                  {step.title}
+                </span>
+                {index < stepData.length - 1 && (
+                  <div className={`ml-4 w-8 h-0.5 ${
+                    currentStep > step.id ? 'bg-amber-600' : 'bg-gray-200'
+                  }`} />
+                )}
+              </div>
+            ))}
+          </div>
+        </div>
+
+        {/* Hauptkarte */}
+        <div className="bg-white rounded-2xl shadow-lg p-8 border border-gray-100">
+          {currentStepData && currentStep <= 3 && (
+            <>
+              {/* Step Header */}
+              <div className="text-center mb-8">
+                <div className={`w-16 h-16 ${currentStepData.iconBg} rounded-full flex items-center justify-center mx-auto mb-4`}>
+                  <span className="text-2xl">{currentStepData.icon}</span>
+                </div>
+                <h2 className="text-2xl font-bold text-gray-900 mb-2">
+                  {currentStepData.subtitle}
+                </h2>
+                <p className="text-gray-600">
+                  {currentStepData.description}
+                </p>
+              </div>
+
+              {/* Formular-Felder */}
+              <div className="space-y-6">
+                <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+                  {currentStepData.fields.map((field) => (
+                    <div key={field.name} className={field.fullWidth ? "md:col-span-2" : ""}>
+                      <label htmlFor={field.name} className="block font-medium mb-2 text-gray-700">
+                        {field.label}
+                        {field.required && <span className="text-red-600 ml-1">*</span>}
+                      </label>
+                      {field.type === "select" ? (
+                        <select
+                          id={field.name}
+                          name={field.name}
+                          value={form[field.name as keyof FormState]}
+                          onChange={handleChange}
+                          className={`w-full p-4 border rounded-xl transition-all ${
+                            errors[field.name] ? "border-red-500 bg-red-50" : "border-gray-300 hover:border-gray-400"
+                          } focus:border-amber-500 focus:outline-none focus:ring-2 focus:ring-amber-200`}
+                        >
+                          <option value="">Bitte wählen</option>
+                          {field.options?.map((option) => (
+                            <option key={option} value={option}>
+                              {option}
+                            </option>
+                          ))}
+                        </select>
+                      ) : (
+                        <input
+                          id={field.name}
+                          name={field.name}
+                          type={field.type || "text"}
+                          value={form[field.name as keyof FormState]}
+                          onChange={handleChange}
+                          autoComplete="off"
+                          inputMode={field.type === "number" ? "numeric" : undefined}
+                          placeholder={field.placeholder}
+                          className={`w-full p-4 border rounded-xl transition-all ${
+                            errors[field.name] ? "border-red-500 bg-red-50" : "border-gray-300 hover:border-gray-400"
+                          } focus:border-amber-500 focus:outline-none focus:ring-2 focus:ring-amber-200`}
+                        />
+                      )}
+                      {errors[field.name] && (
+                        <p className="mt-2 text-red-600 text-sm flex items-center gap-1">
+                          <span className="text-red-500">⚠️</span>
+                          {errors[field.name]}
+                        </p>
+                      )}
+                    </div>
+                  ))}
+                </div>
+              </div>
+
+              {/* Navigation Buttons */}
+              <div className="flex justify-between items-center mt-8 pt-6 border-t border-gray-100">
+                <button
+                  type="button"
+                  onClick={prevStep}
+                  disabled={currentStep === 1}
+                  className={`flex items-center gap-2 px-6 py-3 rounded-xl font-medium transition-all ${
+                    currentStep === 1
+                      ? "text-gray-400 cursor-not-allowed"
+                      : "text-gray-600 hover:text-gray-800 hover:bg-gray-100"
+                  }`}
+                >
+                  <ArrowLeft className="w-4 h-4" />
+                  Zurück
+                </button>
+
+                <button
+                  type="button"
+                  onClick={nextStep}
+                  className="flex items-center gap-2 bg-amber-600 hover:bg-amber-700 text-white px-8 py-3 rounded-xl font-semibold transition-all shadow-md hover:shadow-lg"
+                >
+                  Weiter
+                  <ArrowRight className="w-4 h-4" />
+                </button>
+              </div>
+            </>
           )}
 
-          {errors.form && (
-            <p className="text-red-600 font-medium text-base text-center">
-              {errors.form}
-            </p>
+          {/* Bezahlung-Step */}
+          {currentStep === 4 && (
+            <form onSubmit={handleSubmit}>
+              <div className="text-center mb-8">
+                <div className="w-16 h-16 bg-purple-100 rounded-full flex items-center justify-center mx-auto mb-4">
+                  <span className="text-2xl">💳</span>
+                </div>
+                <h2 className="text-2xl font-bold text-gray-900 mb-2">
+                  Analyse starten
+                </h2>
+                <p className="text-gray-600">
+                  Nur noch ein Klick zur professionellen Pferdebewertung
+                </p>
+              </div>
+
+
+              {/* Einverständnis */}
+              <div className="bg-blue-50 border border-blue-200 rounded-xl p-4 mb-6">
+                <label className="flex items-start gap-3 cursor-pointer">
+                  <input
+                    type="checkbox"
+                    checked={consent}
+                    onChange={(e) => setConsent(e.target.checked)}
+                    className="mt-1 w-5 h-5 text-amber-600 border-gray-300 rounded focus:ring-amber-500"
+                    required
+                  />
+                  <span className="text-sm text-gray-700 leading-relaxed">
+                    Ich stimme ausdrücklich zu, dass die Analyse sofort beginnt, und bestätige mein Erlöschen des Widerrufsrechts gemäß § 356 Abs. 5 BGB.
+                  </span>
+                </label>
+              </div>
+
+              {/* Fehleranzeige */}
+              {errors.form && (
+                <div className="bg-red-50 border border-red-200 rounded-xl p-4 mb-6">
+                  <p className="text-red-700 font-medium text-center">
+                    {errors.form}
+                  </p>
+                </div>
+              )}
+
+              {/* Preis */}
+              <div className="text-center mb-6">
+                <p className="text-lg text-gray-700">
+                  Die Analyse kostet einmalig <strong className="text-amber-600 text-xl">4,90 €</strong>
+                </p>
+                <p className="text-sm text-gray-500">(umsatzsteuerfrei nach § 19 UStG)</p>
+              </div>
+
+              {/* Submit Button */}
+              <button
+                type="submit"
+                disabled={loading || !consent}
+                className="w-full bg-gradient-to-r from-amber-600 to-orange-600 hover:from-amber-700 hover:to-orange-700 disabled:from-gray-300 disabled:to-gray-400 text-white py-4 rounded-xl font-bold text-lg shadow-lg hover:shadow-xl transition-all disabled:cursor-not-allowed"
+              >
+                {loading ? (
+                  <div className="flex items-center justify-center gap-3">
+                    <div className="w-5 h-5 border-2 border-white border-t-transparent rounded-full animate-spin"></div>
+                    Einen Moment – deine Analyse wird vorbereitet…
+                  </div>
+                ) : (
+                  "Jetzt kostenpflichtig analysieren"
+                )}
+              </button>
+
+              {loading && (
+                <p className="text-sm text-gray-500 text-center mt-4">
+                  Bitte einen Moment Geduld – du wirst zur sicheren Zahlung bei Stripe weitergeleitet…
+                </p>
+              )}
+
+              {/* Legal Info */}
+              <div className="text-center mt-6 space-y-2">
+                <p className="text-xs text-gray-500">
+                  Du wirst zur sicheren Bezahlung weitergeleitet.
+                </p>
+                <p className="text-xs text-gray-500">
+                  Mit Klick auf „Jetzt kostenpflichtig analysieren" akzeptierst du unsere{" "}
+                  <Link href="/agb" className="underline hover:text-gray-700">
+                    AGB
+                  </Link>.
+                </p>
+              </div>
+
+              {/* Navigation */}
+              <div className="flex justify-start mt-8 pt-6 border-t border-gray-100">
+                <button
+                  type="button"
+                  onClick={prevStep}
+                  className="flex items-center gap-2 text-gray-600 hover:text-gray-800 hover:bg-gray-100 px-6 py-3 rounded-xl font-medium transition-all"
+                >
+                  <ArrowLeft className="w-4 h-4" />
+                  Zurück
+                </button>
+              </div>
+            </form>
           )}
+        </div>
+      </section>
 
-          <p className="text-sm text-gray-600 text-center mb-4">
-            Die Analyse kostet einmalig <strong>4,90 €</strong> (umsatzsteuerfrei nach § 19 UStG)
-          </p>
+      {/* Social Proof */}
+      <section className="bg-white py-16 px-4 -mx-4 mt-16">
+        <div className="max-w-6xl mx-auto">
+          <h2 className="text-3xl font-bold text-gray-800 text-center mb-12">
+            Das sagen unsere Kunden
+          </h2>
+          
+          <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
+            {/* Testimonial 1 */}
+            <div className="bg-gradient-to-br from-white to-gray-50 rounded-xl p-6 shadow-md border border-gray-100">
+              <div className="flex mb-4">
+                {[...Array(5)].map((_, i) => (
+                  <Star key={i} className="w-5 h-5 text-yellow-400 fill-current" />
+                ))}
+              </div>
+              <blockquote className="text-gray-700 mb-4 leading-relaxed">
+                "Ich wollte mein Pferd verkaufen und war unsicher beim Preis. Die Bewertung hat mir sehr geholfen eine Einschätzung zu bekommen und ich konnte mein Pferd auch zu dem empfohlenen Preis verkaufen!"
+              </blockquote>
+              <cite className="text-sm text-gray-600 font-semibold not-italic">
+                - Sarah M., Freizeitreiterin
+              </cite>
+            </div>
 
-          <button
-            type="submit"
-            disabled={loading}
-            className="w-full bg-brand-accent text-white py-4 rounded-2xl font-bold text-button shadow-soft hover:bg-brand transition"
-          >
-            {loading ? (
-              <>
-                <span className="inline-block animate-spin mr-2 h-5 w-5 border-2 border-white border-t-transparent rounded-full"></span>
-                Einen Moment – deine Analyse wird vorbereitet…
-              </>
-            ) : (
-              "Jetzt kostenpflichtig analysieren"
-            )}
-          </button>
+            {/* Testimonial 2 */}
+            <div className="bg-gradient-to-br from-white to-gray-50 rounded-xl p-6 shadow-md border border-gray-100">
+              <div className="flex mb-4">
+                {[...Array(5)].map((_, i) => (
+                  <Star key={i} className="w-5 h-5 text-yellow-400 fill-current" />
+                ))}
+              </div>
+              <blockquote className="text-gray-700 mb-4 leading-relaxed">
+                "Vor dem Pferdekauf wollte ich wissen, ob der angegebene Preis fair ist. Die PferdeWert-Analyse war sehr detailliert und hat mir bei der Preisverhandlung sehr geholfen."
+              </blockquote>
+              <cite className="text-sm text-gray-600 font-semibold not-italic">
+                - Michael K., Hobbyreiter
+              </cite>
+            </div>
 
-          {loading && (
-            <p className="text-sm text-gray-500 text-center mt-4">
-              Bitte einen Moment Geduld – du wirst zur sicheren Zahlung bei Stripe weitergeleitet…
-            </p>
-          )}
-
-          <p className="text-xs text-gray-500 text-center mt-2">
-            Du wirst zur sicheren Bezahlung weitergeleitet.
-          </p>
-          <p className="text-xs text-gray-500 text-center">
-            Mit Klick auf „Jetzt kostenpflichtig analysieren“ akzeptierst du unsere <Link href="/agb" className="underline">AGB</Link>.
-          </p>
-        </form>
-      </main>
+            {/* Testimonial 3 */}
+            <div className="bg-gradient-to-br from-white to-gray-50 rounded-xl p-6 shadow-md border border-gray-100">
+              <div className="flex mb-4">
+                {[...Array(5)].map((_, i) => (
+                  <Star key={i} className="w-5 h-5 text-yellow-400 fill-current" />
+                ))}
+              </div>
+              <blockquote className="text-gray-700 mb-4 leading-relaxed">
+                "Ich besitze ein Pferd und wollte einfach nur aus Neugier den aktuellen Marktwert wissen. Super interessant was PferdeWert als Ergebnis bereitstellt, vor allem auch die Analyse der Abstammung fand ich sehr spannend!"
+              </blockquote>
+              <cite className="text-sm text-gray-600 font-semibold not-italic">
+                - Anna L., Pferdebesitzerin
+              </cite>
+            </div>
+          </div>
+        </div>
+      </section>
     </Layout>
   );
 }
