+++ frontend/tsconfig.json
@@ -18,6 +18,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "../archive/bewerten-alt.tsx"],
   "exclude": ["node_modules"]
 }
